Multi_PDF_Doc_Langchain_Gemini/requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
pandas 
openpyxl 
python-docx
faiss-cpu
langchain_google_genai